requirements.txt
aiohttp==3.6.3
antiorm==1.2.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
discord==1.0.1
discord.py==1.5.1
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
html5lib==1.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.1.1
jdcal==1.4.1
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.3
mpmath==1.1.0
multidict==4.7.6
mysql-connector-python==8.0.23
numpy==1.19.4
opencv-python==4.4.0.46
openpyxl==3.0.5
pandas==1.2.0
parse==1.19.0
Pillow==8.0.1
psycopg2==2.8.6
pyaes==1.6.1
pycparser==2.20
pyee==8.1.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
Pyrogram==1.1.13
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
sympy==1.7
tornado==6.1
tqdm==4.56.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
yarl==1.5.1
zipp==3.4.0
